Secondary popout window, Corrected Os Cursos visibility, adjusted naming
Added ability to toggle item spawner and have it pop out on the right side of the window.

Added ability to enable and disable the correct OS Cursor without flickering!

Rearranged the buttons for item spawner so that it starts with common and goes up to immortal, readjusted the coloring to be more understandable